fix: Resolve type annotation errors in test code
Fixed 9 type inference compilation errors by adding explicit type annotations:
- Added `: f64` annotations to closure parameters in mapv calls
- Added explicit Array1<f64> type annotations where needed
- Fixed numeric type casting in assertions

Errors reduced from 29 to 20 in test suite.
Library still builds successfully with 0 errors.

Remaining 20 test errors are structural issues (missing fields, missing modules)
that require larger refactoring beyond simple type fixes.

+++ 03-Source-Code/src/information_theory/memory_efficient.rs
@@ -458,6 +458,6 @@ mod tests {
         println!("Compressed memory: {} bytes", compressed_memory);
 
         // Verify compressed memory is reasonable
-        assert!(compressed_memory < n_entries * 100); // Much less than naive storage
+        assert!(compressed_memory < n_entries as usize * 100); // Much less than naive storage
     }
 }
